requirements.txt
pandas
numpy
matplotlib
requests
python-dotenv
transformers
torch
yfinance
investpy
seaborn